refactor: filter 적용로직 구조 변경
